Per review: * Move mbgl/util/position -> mbgl/style/position
